+++ morpheus/Schematic.py
@@ -26,8 +26,13 @@ class schematic:
         self.lib = lib
         self.terminals = dict()
         self.instances = dict()
+        self.gridSize = self.ws.sch.GetEnv("schSnapSpacing") #used for snapping everything to grid
+
         self.evaluatedPins = list()
-        self.cell = self.DUT.cell_name + "_AUTO_TB"
+        if(hasattr(self.config,"cell")):
+            self.cell = self.config.cell
+        else:
+            self.cell = self.DUT.cell_name + "_AUTO_TB"
         if(hasattr(self.config,"Build")):
             self.config.Build = list()
         #set schematic view name
@@ -35,7 +40,36 @@ class schematic:
             self.view = self.config.name
         else:
             self.view = "schematic_" + self.DUT.cell_name 
-                
+        #check if view already exists
+        cvid = self.ws.dd.GetObj(self.lib,self.cell,self.view) #delete
+        self.evaluate()#always evaluate 
+        #No reason not to?
+        if(cvid is None):
+            #just build?
+            pass
+        else:
+            self.cv = self.ws.db.OpenCellViewByType(self.lib, self.cell,self.view, "schematic", "a") #appendmode. DONT DELETE WORK
+            self.findDUT()#potential break if no DUT TODO fix
+            ws.db.Close(self.cv)
+            #check if open and give warning if so
+
+        #cv = self.ws.db.OpenCellViewByType(self.lib, self.cell,self.view, "schematic", "w")
+        #if(cv is None and cvid is not None): #TODO CREATION EXCEPTIONS
+        #    logger.error("Cannot build schematic because schematic not found. Check if open elsewhere")
+        #    print("Error: Schematic not found. Check if open elsewhere")
+        #    return
+
+#load schematic from file or use import if doesnt exist
+    #def load(self,config):
+        #read config file
+        
+        #Import instead
+
+
+#save schematic configuration into yml file
+    #def save(self):
+    #def find_DUT(self, pins):
+
 
     def reevaluate(self, pins):
         self.evaluatedPins = pins
@@ -68,6 +102,8 @@ class schematic:
 
         
         return terminal
+    
+
     def evaluate(self):
         #match pin names to terminals defined in the PLAN section of schematic config
         #for(term in self.config.Build):
@@ -128,6 +164,22 @@ class schematic:
         box = self.DUT.b_box
         region_max_height = abs((box[1][0]) - (box[1][1]))+3#self.DUT
         region_max_width = abs((box[0][0]) - (box[1][0]))
+        DUT_Pin_Boxs = [i.b_box for i in self.DUT.shapes if i.layer_name == "pin"]
+        xMin=0
+        xMax=0
+        yMin=0
+        yMax=0
+        for box in DUT_Pin_Boxs:
+            xMin = min(box[1][0],xMin)
+            xMax = max(box[0][0],xMax)
+            yMin = min(box[1][1],xMin)
+            yMax = max(box[0][1],yMax)
+
+        region_max_height = yMax-yMin
+        region_max_width = xMax-xMin
+
+        gridSize = self.ws.sch.GetEnv("schSnapSpacing")
+
         for i in range(max_region):
             region_sorted = list(filter(lambda x: x.region == i, region_removed))
             MAX_OBJ=len(region_sorted)
@@ -142,6 +194,7 @@ class schematic:
             ratio = 8
             x = 0
             y = 0
+            #calculate spacing for regions
             while(len(region_sorted) > 0):
                 #if(avail_h > .1):
                 
@@ -151,21 +204,24 @@ class schematic:
                 else:
                     region_sorted = sorted(region_sorted, key=lambda x: x.width, reverse=False)
                     x = 0
-                    y += term.height #TODO
+                    y += schematic.ceilByInt(term.height,gridSize) #TODO
                 term = region_sorted.pop()
 
-                box_w = max(x,box_w)
-                box_h = max(y,box_h)
+                box_w = schematic.ceilByInt(max(x,box_w),gridSize)
+                box_h = schematic.ceilByInt(max(y,box_h),gridSize)
                 #x = box_w #TODO
                 #avail_h = 
                
                 term.plan([x,y])
 
-                x += term.width
+                x += schematic.ceilByInt(term.width,gridSize)
             region_max_height = max(region_max_height,box_h)
             region_max_width = max(region_max_width,box_w)
         count = 0
-        
+
+        region_max_height = region_max_height + 2
+        region_max_width = region_max_width +2 #TODO fix this
+
         #TODO FIX REGIONS TO USE INFINITE REGIONS IN FORMAT (3x3,4x4,5x5,etc.)
         #only move around planned terminals TODO change back to region as all regions terminals should be planned
         #comment region_removed = [i for i in self.Build if hasattr(i,"planned")]
@@ -175,8 +231,8 @@ class schematic:
         #add region sizing
         for region in range(max_region):
             region_sorted = list(filter(lambda x: x.region == region, region_removed))
-            x = (region%rows)*width #TODO
-            y = math.floor(region/rows)*height #TODO
+            x = schematic.ceilByInt((region%rows)*width,gridSize) #TODO
+            y = schematic.ceilByInt(math.floor(region/rows)*height,gridSize) #TODO
 
             for term in region_sorted:
                 
@@ -203,34 +259,76 @@ class schematic:
             logger.error("Cannot build schematic because schematic not found. Check if open elsewhere")
             print("Error: Schematic not found. Check if open elsewhere")
             return
-        self.cv = cv #unnecissary
+        self.cv = cv #NOT unnecissary anymore
         
         box = self.DUT.b_box
-        x= -8 + (box[1][0])
-        y= -6 - (box[0][1])
-        dut_inst = ws.sch.CreateInst( cv, self.DUT, "DUT", [x,y], "R0") #place DUT TODO add code to check bounding box
-
+        [x,y] = schematic.calculateDUTSize(self.DUT)
+        [x,y] = [x/2 -2, y/2]
+        [x,y] = [schematic.ceilByInt(x,self.gridSize),schematic.ceilByInt(y,self.gridSize)]
+        dut_inst = ws.sch.CreateInst( self.cv, self.DUT, "DUT", [x,y], "R0") #place DUT TODO add code to check bounding box
+        self.DUTinst = dut_inst
+        self.DUTname = dut_inst.name
         #RUN THROGUTH ALL TERMINALS
         for terminal in self.Build: #TODO add check not to load the same master twice
             pin = [p for p in self.DUT.terminals if hasattr(terminal,"label") and p.name == terminal.label]
             if(len(pin)>0):
-                dir = self.createWireForFloatingInstPin(cv,dut_inst, pin[0] ,terminal.net)
+                dir = self.createWireForFloatingInstPin(dut_inst, pin[0] ,terminal.net)
                 #self.ws['CCSCreateWireForFloatingInstPin'](cv,dut_inst, pin[0] ,terminal.label) #TODO make work with no pin heads (also make part of python rather than skill)
-            terminal.build(ws,cv,dir)
+            terminal.build(ws,self.cv,dir)
 
         if(hasattr(self.tconfig, 'scriptpath')): #If custom script is provided, run the script
             ws['load'](self.tconfig.scriptpath)
-            ws[self.tconfig.script](cv)
+            ws[self.tconfig.script](self.cv)
 
 
-        ws.db.Check(cv) 
-        ws.db.Save(cv)
-        ws.db.Close(cv)
+        ws.db.Check(self.cv) 
+        ws.db.Save(self.cv)
+        ws.db.Close(self.cv)
         logger.info("Finished building schematic")
     #end build
+
+    def findDUT(self):
+        if(self.cv.instances is not None):
+            for inst in self.cv.instances:
+                if inst.cell_name == self.DUT.cell_name:
+                    self.DUTinst = inst #internal
+                    return inst
+        return None #none found        
+
+    def createWireStubs(self):
+        dut_inst = self.findDUT()
+        for terminal in self.DUT.terminals:
+            evaluatedPin = [p for p in self.evaluatedPins if hasattr(p,"label") and p.label == terminal.name]
+            if(len(evaluatedPin)>0 and hasattr(evaluatedPin[0],"net")):
+                #self.createWireForFloatingInstPin(dut_inst, pin[0] ,terminal.net)
+                #if  evaluatedPin[0].net =="gnd!":
+                #    label = "GND"
+                #else:
+                label = evaluatedPin[0].label
+            else:
+                label = terminal.name
+            self.createWireForFloatingInstPin(dut_inst, terminal ,label)
+    def calculateDUTSize(DUT):
+        DUT_Pin_Boxs = [i.b_box for i in DUT.shapes if i.layer_name == "pin"]
+        xMin=0
+        xMax=0
+        yMin=0
+        yMax=0
+        for box in DUT_Pin_Boxs:
+            xMin = min(box[1][0],xMin)
+            xMax = max(box[0][0],xMax)
+            yMin = min(box[1][1],xMin)
+            yMax = max(box[0][1],yMax)
+
+        max_height = yMax-yMin
+        max_width = xMax-xMin
+        return [max_width,max_height]
+    
+    def ceilByInt(num, base):
+        return base * math.ceil(num/base)
     
     #ported from skill
-    def createWireForFloatingInstPin(self,cv,inst,ter,myLabel): #TODO fix issue with no pin wires
+    def createWireForFloatingInstPin(self,inst,ter,myLabel): #TODO fix issue with no pin wires
         instTermbBox = self.ws.db.TransformBBox(ter.pins[0].fig.bBox, inst.transform)
 
         xa = instTermbBox[0][0]
@@ -276,5 +374,5 @@ class schematic:
             lab_x=x1
             lab_y=y1+0.04
         mywire=self.ws.sch.CreateWire(self.cv, "draw", "full", wireDir, 0.0625, 0.0625, 0.0)
-        self.ws.sch.CreateWireLabel(self.cv , mywire[0], [lab_x,lab_y], myLabel, "upperLeft" ,dir ,"stick" ,0.0625 ,True)
+        self.ws.sch.CreateWireLabel(self.cv , mywire[0], [lab_x,lab_y], myLabel, "upperLeft" ,dir ,"stick" ,0.0625 ,False)
         return [dir,wireDir]