+++ morpheus/Terminal.py
@@ -1,5 +1,6 @@
 from morpheus.Config import config
 from morpheus.Config import config_types
+import re
 
 class Terminal:
     terminals = dict()
@@ -27,10 +28,12 @@ class Terminal:
         self.width = self.config.width
         self.height = self.config.height
         self.eval = True
-        if(self.type == 'gnd'):
-            self.net = 'gnd!'
+        #if(self.type == 'gnd'):
+        #    self.net = 'gnd!'
         if(hasattr(self.term, "label")):
             self.net = self.term.label
+        #if(hasattr(self.term, "net")):
+        #    self.net = self.term.net
 
     def plan(self,position):
         self.position = position
@@ -45,7 +48,7 @@ class Terminal:
 
     def build(self,ws,cv,dir):
         #TODO add special case for no connect
-        if(self.type == "NOCONNECT"):
+        if(self.term.term == "NOCONNECT"):
             name = self.label + "_NOCONN"
             rt =ws.db.ConcatTransform([[0,0],dir[0]],[[0,0],"R90"])
             instance = ws.db.OpenCellViewByType("basic", "noConn", "symbol")
@@ -63,7 +66,7 @@ class Terminal:
                 if(inst.name == "wire"): #WIRE SPECIAL CASE
                     mywire = ws.sch.CreateWire(cv, "draw", "full", position, 0.0625, 0.0625, 0.0)
                     if(hasattr(inst,"label") and hasattr(self,"label")):
-                        ws.sch.CreateWireLabel(cv, mywire[0], [position[0][0] + 0.0325,position[0][1] + 0.05], inst.label.format(**self.dictionary), "upperLeft", "R90", "stick", 0.0625, True)
+                        ws.sch.CreateWireLabel(cv, mywire[0], [position[0][0] + 0.0325,position[0][1] + 0.05], inst.label.format(**self.dictionary), "upperLeft", "R90", "stick", 0.0625, False)
                 else:
                     symbol = Terminal.getInst(ws,inst)
                     name = self.config.name + "_" + inst.name #TODO BETTER NAMING SYSTEM
@@ -79,10 +82,15 @@ class Terminal:
                         if(hasattr(self.term,"parameters")): #Terminal takes priority over default
                             for param in self.term.parameters.__dict__:
                                 #TODO ERROR CATCH NUMBERS
-                                property = self.term.parameters.__dict__.get(param).format(name = self.label)
+                                propText = self.term.parameters.__dict__.get(param)
+                                property = propText.format(name = self.label)
+                                if propText.find("regex") != -1:
+                                    property = re.sub(self.term.pattern,propText.replace("regex","") ,self.label) #Do regular expression replace
                                 prop.append([param,"string",property])
-                                    
-                    new_inst = ws.db.CreateParamInst(cv, symbol, name, position, "R0",1,prop) #create instance with parameters
+                    rotation = "R0" #add rotate functionality
+                    if(hasattr(inst,"rotation")):
+                        rotation = "R" + str(inst.rotation)   
+                    new_inst = ws.db.CreateParamInst(cv, symbol, name, position, rotation,1,prop) #create instance with parameters
             #end For Loop per Instance
         #end For Loop per Terminal
     def getTerminal(term_type):